+++ include/core/math/vector/vector3i.hpp
@@ -9,10 +9,10 @@ namespace core::math {
 #pragma warning(disable: 4201)
 	struct Vector3i {
 		union {
-			struct { s64 x, y, z; };
-			struct { s64 w, h, l; };
-			struct { s64 r, g, b; };
-			s64 data[3];
+			struct { s32 x, y, z; };
+			struct { s32 w, h, l; };
+			struct { s32 r, g, b; };
+			s32 data[3];
 		};
 	};
 #pragma warning(pop)
@@ -29,8 +29,8 @@ namespace core::math {
 	constexpr Vector3i operator-(Vector3i v1, Vector3i v2) { return { (v1.x - v2.x), (v1.y - v2.y), (v1.z - v2.z) }; }
 	constexpr Vector3i operator*(Vector3i v1, Vector3i v2) { return { (v1.x * v2.x), (v1.y * v2.y), (v1.z * v2.z) }; }
 
-	constexpr Vector3i operator*(Vector3i v, s64 scaler) { return { (v.x * scaler), (v.y * scaler), (v.z * scaler) }; }
-	constexpr Vector3i operator*(s64 scaler, Vector3i v) { return v * scaler; }
+	constexpr Vector3i operator*(Vector3i v, s32 scaler) { return { (v.x * scaler), (v.y * scaler), (v.z * scaler) }; }
+	constexpr Vector3i operator*(s32 scaler, Vector3i v) { return v * scaler; }
 
 	constexpr bool operator==(Vector3i v1, Vector3i v2) { return (v1.x == v2.x) && (v1.y == v2.y) && (v1.z == v2.z); }
 	constexpr bool operator!=(Vector3i v1, Vector3i v2) { return (v1.x != v2.x) || (v1.y != v2.y) && (v1.z != v2.z); }
